Remove credentials
