+++ demo/pom.xml
@@ -25,7 +25,7 @@
         </dependency>
         <dependency>
             <groupId>com.ibm.cloud</groupId>
-            <artifactId>data-cfenv-processor</artifactId>
+            <artifactId>amqp-cfenv-processor</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
         <dependency>
