+++ demo/src/main/java/com/example/demo/amqp/Receiver.java
@@ -13,7 +13,7 @@
  *   See the License for the specific language governing permissions and
  *   limitations under the License.
  */
-package com.example.demo;
+package com.example.demo.amqp;
 
 import org.springframework.stereotype.Component;
